+++ EmpresaTransporte/nbproject/private/private.properties
@@ -1,3 +1,4 @@
+<<<<<<< Updated upstream
 <<<<<<< HEAD
 compile.on.save=true
 user.properties.file=/Users/lennonsantos/Library/Application Support/NetBeans/8.1/build.properties
@@ -5,3 +6,12 @@ user.properties.file=/Users/lennonsantos/Library/Application Support/NetBeans/8.
 compile.on.save=true
 user.properties.file=C:\\Users\\Gustavo\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
 >>>>>>> 6a6b4a9a56e8f3acb02d067dd080e0caa22801c4
+=======
+<<<<<<< HEAD
+compile.on.save=true
+user.properties.file=/Users/lennonsantos/Library/Application Support/NetBeans/8.1/build.properties
+=======
+compile.on.save=true
+user.properties.file=/Users/lennonsantos/Library/Application Support/NetBeans/8.1/build.properties
+>>>>>>> 6a6b4a9a56e8f3acb02d067dd080e0caa22801c4
+>>>>>>> Stashed changes
