+++ EmpresaTransporte/src/empresatransporte/Caminhao.java
@@ -1,5 +1,7 @@
 package empresatransporte;
 
-public class Caminhao {
-    
+public class Caminhao extends Veiculo{
+    public Caminhao(String mod, String mot, int anoF, int anoM, String cor, String com, double val){
+      super(mod, mot, anoF, anoM, cor, com, val);
+    }
 }
