+++ EmpresaTransporte/src/empresatransporte/Carro.java
@@ -1,5 +1,9 @@
 package empresatransporte;
 
-public class Carro {
+public class Carro extends Veiculo{
+    
+    public Carro(String mod, String mot, int anoF, int anoM, String cor, String com, double val){
+        super(mod, mot, anoF, anoM, cor, com, val);
+    }
     
 }
