+++ EmpresaTransporte/src/empresatransporte/EmpresaTransporte.java
@@ -3,7 +3,8 @@ package empresatransporte;
 public class EmpresaTransporte {
 
     public static void main(String[] args) {
-
+        Tela t = new Tela();
+        t.setVisible(true);
     }
     
 }
